[wagtail] Refactor pagination into a helper
Pagination is done almost exactly the same across all of Wagtail. Moving
the repeated code in to one spot ensures mistakes are not made, and
means that pagination is always done consistently.

There are some locations where the pagination helper has not been used,
as there is no request available or something similar. These instances
are special enough that refactoring them does not make sense at this
point.

Conflicts:
	wagtail/wagtailadmin/views/chooser.py
	wagtail/wagtailadmin/views/pages.py
	wagtail/wagtaildocs/views/chooser.py
	wagtail/wagtaildocs/views/documents.py
	wagtail/wagtailforms/views.py
	wagtail/wagtailimages/views/chooser.py
	wagtail/wagtailimages/views/images.py
	wagtail/wagtailredirects/views.py
	wagtail/wagtailsearch/views/editorspicks.py
	wagtail/wagtailsnippets/views/chooser.py
	wagtail/wagtailsnippets/views/snippets.py
	wagtail/wagtailusers/views/groups.py
	wagtail/wagtailusers/views/users.py