[wagtail] Use six from Django, remove six from tox.ini
All referencs to `six` were removed in #1411 favouring
`django.utils.six`, except for one in `tox.ini` where it was installed
as a dependency. Because it was still installed in the test environment,
a few references to `six` crept back in. These will cause an ImportError
when users do not have `six` installed.

These references have been fixed, and `six` has been removed from
`tox.ini`