[wagtail] Disable client-side validation on all admin forms #2942
* Revert "Disable client-side validation on Django 1.10 for multipart forms"

This reverts commit 197d85ce9f674e319cc7d7d0fa6e0a05dc6a46d7.

* Disable client-side validation on all admin forms

Client-side validation fails to handle all the nuances of how forms are
used in Wagtail, such as indicating errors across tabbed forms, and generally
results in a worse user experience than our existing server-side validation.

Fixes #2934

* Don't pre-emptively validate long-running forms if novalidate is set

We run checkValidity before enabling the spinner on long-running forms, so that we don't end up with a form that's been blocked from submission but still has the spinner running. Unfortunately, if novalidate is set, then checkValidity still returns False and aborts enabling the spinner, even though the 'real' validation (the one that prevents submission) never happens.

To make matters worse, if we abort before we've set up the cancelSpinner function, and then promptly encounter a custom onsubmit handler that recognises the error and attempts to call cancelSpinner, the JS breaks completely.

Eww.

To remedy this, we now only run checkValidity on forms that don't have novalidate, and only after setting up cancelSpinner.