[wagtail] Revise wagtailsearch migrations to avoid keeping editorspick table around
Fixes #1824, as per the outline in https://github.com/wagtail/wagtail/issues/1824#issuecomment-1271840741. Revise the existing migrations so that rather than keeping the wagtailsearch_editorspick table in place and unmanaged for the wagtail.contrib.search_promotions app to potentially rename and adopt, search_promotions creates its own table. This then leaves wagtailsearch free to delete wagtailsearch_editorspick properly in a new migration, to be run on both new and upgraded projects (although this needs to be done inside an existence check, in case the old version of the search_promotions app migrations have already run and renamed the table).