[wagtail] Fix search test migration dependency for older taggit versions
The searchtests.0002_bookunindexed migration had a dependency on taggit migration 0005, which was added in django-taggit 3.0, but Wagtail only specifies django-taggit >=2.0, so this fails on older django-taggit versions. We don't want to bump this yet (since people could still legitimately be using django-taggit 2.x against Django 3.2 LTS), so fix this migration to work against older django-taggit releases.