[wagtail] Remove fallback for fields that don't exist in search_fields
The `check` method in BaseSearchQueryCompiler rejects these, so this doesn't make a difference to real-world queries. However, the check is only performed after compilation is complete, so we still need to guard against this with `searchable_fields.get`.

However, our tests _do_ blithely use non-existent fields, so fix those too...